+++ pkg/codegen/internal/test/providers.go
@@ -1,0 +1,25 @@
+package test
+
+import (
+	"io/ioutil"
+	"path/filepath"
+
+	"github.com/pulumi/pulumi/pkg/resource/deploy/deploytest"
+	"github.com/pulumi/pulumi/sdk/go/common/resource/plugin"
+)
+
+func GetSchema(schemaDirectoryPath, providerName string) ([]byte, error) {
+	return ioutil.ReadFile(filepath.Join(schemaDirectoryPath, providerName+".json"))
+}
+
+func AWS(schemaDirectoryPath string) (plugin.Provider, error) {
+	schema, err := GetSchema(schemaDirectoryPath, "aws")
+	if err != nil {
+		return nil, err
+	}
+	return &deploytest.Provider{
+		GetSchemaF: func(version int) ([]byte, error) {
+			return schema, nil
+		},
+	}, nil
+}