+++ CHANGELOG.md
@@ -26,6 +26,9 @@ CHANGELOG
 - Define merge behavior for resource options in Go SDK
   [#4316](https://github.com/pulumi/pulumi/pull/4316)
 
+- Add overloads to Output.All in .NET
+  [#4321](https://github.com/pulumi/pulumi/pull/4321)
+
 ## 1.14.0 (2020-04-01)
 - Fix error related to side-by-side versions of `@pulumi/pulumi`.
   [#4235](https://github.com/pulumi/pulumi/pull/4235)